buildBinFile uses docker
Embed directories are using tar gz to embed files,
in order to keep the same tar.gz fingerprint,
we need to have a consistent environment.
So the compilation is done using docker image
scrasnups/build:bash-tools-ubuntu-5.3.
Also probably the difference comes from the
ownership that is written in the archive, we now
uses these options with tar command:
 --sort=name --mtime="@0" \
--owner=0 --group=0 --numeric-owner \
--pax-option=exthdr.name=%d/PaxHeaders/%f,delete=atime,delete=ctime,delete=mtime \
--no-same-owner